Latest Release RM1103-0316-0.93.1-f5756a0 on PATREON - UPD TOTP

+++ applications/external/totp/lib/wolfssl/wolfcrypt/src/md2.c
@@ -107,7 +107,7 @@ void wc_Md2Update(Md2* md2, const byte* data, word32 len)
                     t = md2->X[j+6] ^= S[t];
                     t = md2->X[j+7] ^= S[t];
                 }
-                t = (t + i) & 0xFF;
+                t = (byte)((t + i) & 0xFF);
             }
         }
     }
